test(signing): split testsigning fixtures into per-method files
